1.对外服务:嵌入网站SDK支持提供初始化参数
2.数据表integer,number类型字段数据存储与渲染优化
3.修复机器人提示词内输入/插入变量,变量列表无法下拉
4.对外服务嵌入网站增加弹窗大小和位置设置
5.Agent查询本地知识库工具支持大模型自主选择查询那个知识库
6.Agent机器人、聊天机器人、工作流增加限流设置
7.【STD】关联的chatwiki机器人:开启多模态时,额外转发图片语音视频消息
8.Agent机器人:关联知识库优化
9.修复限流无回复时,对话聊天测试和对外服务用户消息被聊天遮挡
10.【STD】云版支持钉钉、企微、飞书登录


https://www.yuque.com/zhimaxiaoshiwangluo/pggco1/tvwn5npk63aqikq1

+++ internal/app/chatwiki/work_flow/adapter.go
@@ -1332,7 +1332,7 @@ func (n *FormSelectNode) Running(flow *WorkFlow) (output common.SimpleFields, ne
 				case `integer`:
 					value = cast.ToInt(value)
 				case `number`:
-					value = cast.ToFloat32(value)
+					value = cast.ToFloat64(value)
 				case `boolean`:
 					value = cast.ToBool(value)
 				}
